requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2
python-multipart==0.0.22
langchain-core==1.2.19
langchain-openai==1.1.11
langsmith==0.7.17
openai==2.28.0
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
httpx==0.28.1
requests==2.32.5
rapidfuzz==3.13.0
pydantic==2.12.5
starlette==0.52.1
